refactor: Rename HistoryCleanupMapper to split it later
