+++ db/rdbms/src/main/java/io/camunda/db/rdbms/sql/IncidentMapper.java
@@ -12,7 +12,7 @@ import io.camunda.db.rdbms.write.domain.IncidentDbModel;
 import io.camunda.search.entities.IncidentEntity;
 import java.util.List;
 
-public interface IncidentMapper extends HistoryCleanupMapper {
+public interface IncidentMapper extends ProcessBasedHistoryCleanupMapper {
 
   void insert(IncidentDbModel incident);
