+++ db/rdbms/src/main/java/io/camunda/db/rdbms/sql/JobMapper.java
@@ -11,7 +11,7 @@ import io.camunda.db.rdbms.read.domain.JobDbQuery;
 import io.camunda.db.rdbms.write.domain.JobDbModel;
 import java.util.List;
 
-public interface JobMapper extends HistoryCleanupMapper {
+public interface JobMapper extends ProcessBasedHistoryCleanupMapper {
 
   void insert(JobDbModel job);
