+++ db/rdbms/src/main/java/io/camunda/db/rdbms/sql/VariableMapper.java
@@ -13,7 +13,7 @@ import io.camunda.search.entities.VariableEntity;
 import io.camunda.util.ObjectBuilder;
 import java.util.List;
 
-public interface VariableMapper extends HistoryCleanupMapper {
+public interface VariableMapper extends ProcessBasedHistoryCleanupMapper {
 
   void insert(VariableDbModel variable);
