+++ db/rdbms/src/main/java/io/camunda/db/rdbms/write/service/ProcessInstanceWriter.java
@@ -7,8 +7,8 @@
  */
 package io.camunda.db.rdbms.write.service;
 
-import io.camunda.db.rdbms.sql.HistoryCleanupMapper;
-import io.camunda.db.rdbms.sql.HistoryCleanupMapper.CleanupHistoryDto;
+import io.camunda.db.rdbms.sql.ProcessBasedHistoryCleanupMapper;
+import io.camunda.db.rdbms.sql.ProcessBasedHistoryCleanupMapper.CleanupHistoryDto;
 import io.camunda.db.rdbms.sql.ProcessInstanceMapper;
 import io.camunda.db.rdbms.sql.ProcessInstanceMapper.EndProcessInstanceDto;
 import io.camunda.db.rdbms.write.domain.ProcessInstanceDbModel;
@@ -117,7 +117,7 @@ public class ProcessInstanceWriter {
               WriteStatementType.UPDATE,
               processInstanceKey,
               "io.camunda.db.rdbms.sql.ProcessInstanceMapper.updateHistoryCleanupDate",
-              new HistoryCleanupMapper.UpdateHistoryCleanupDateDto.Builder()
+              new ProcessBasedHistoryCleanupMapper.UpdateHistoryCleanupDateDto.Builder()
                   .processInstanceKey(processInstanceKey)
                   .historyCleanupDate(historyCleanupDate)
                   .build()));
